+++ src/Security/Authentication/Negotiate/Samples/NegotiateAuthSample/NegotiateAuthSample.csproj
