+++ src/Security/Authentication/Negotiate/Samples/NegotiateAuthSample/NtAuthWrapper.cs
@@ -8,7 +8,7 @@ using System.Security.Claims;
 using System.Security.Principal;
 using System.Threading.Tasks;
 
-namespace IntegratedAuthSample
+namespace NegotiateAuthSample
 {
     public class NtAuthWrapper
     {
