+++ src/Security/Authentication/Negotiate/Samples/NegotiateAuthSample/appsettings.Development.json
