+++ src/Security/Authentication/Negotiate/Samples/NegotiateAuthSample/appsettings.json
